@@ -18,17 +18,19 @@
 from typing import Any, Optional, TYPE_CHECKING, Union
 import httpx
 import json
+from . import _common
 
 
 if TYPE_CHECKING:
   from .replay_api_client import ReplayResponse
+  import aiohttp
 
 
 class APIError(Exception):
   """General errors raised by the GenAI API."""
 
   code: int
-    response: Union["ReplayResponse", httpx.Response]
+  response: Union['ReplayResponse', httpx.Response, 'aiohttp.ClientResponse']
 
   status: Optional[str] = None
   message: Optional[str] = None
@@ -37,48 +39,53 @@ class APIError(Exception):
       self,
       code: int,
       response_json: Any,
-        response: Optional[Union["ReplayResponse", httpx.Response]] = None,
+      response: Optional[
+          Union['ReplayResponse', httpx.Response, 'aiohttp.ClientResponse']
+      ] = None,
   ):
+    if isinstance(response_json, list) and len(response_json) == 1:
+      response_json = response_json[0]
+
     self.response = response
     self.details = response_json
     self.message = self._get_message(response_json)
     self.status = self._get_status(response_json)
     self.code = code if code else self._get_code(response_json)
 
-        super().__init__(f"{self.code} {self.status}. {self.details}")
+    super().__init__(f"{self.code} {self.status}. {self.details}")
 
   def _get_status(self, response_json: Any) -> Any:
     return response_json.get(
-            "status", response_json.get("error", {}).get("status", None)
+        "status", response_json.get("error", {}).get("status", None)
     )
 
   def _get_message(self, response_json: Any) -> Any:
     return response_json.get(
-            "message", response_json.get("error", {}).get("message", None)
+        "message", response_json.get("error", {}).get("message", None)
     )
 
   def _get_code(self, response_json: Any) -> Any:
     return response_json.get(
-            "code", response_json.get("error", {}).get("code", None)
+        "code", response_json.get("error", {}).get("code", None)
     )
 
-  def _to_replay_record(self) -> dict[str, Any]:
+  def _to_replay_record(self) -> _common.StringDict:
     """Returns a dictionary representation of the error for replay recording.
 
     details is not included since it may expose internal information in the
     replay file.
     """
     return {
-            "error": {
-                "code": self.code,
-                "message": self.message,
-                "status": self.status,
+        "error": {
+            "code": self.code,
+            "message": self.message,
+            "status": self.status,
         }
     }
 
   @classmethod
   def raise_for_response(
-        cls, response: Union["ReplayResponse", httpx.Response]
+      cls, response: Union["ReplayResponse", httpx.Response]
   ) -> None:
     """Raises an error with detailed error message if the response has an error status."""
     if response.status_code == 200:
@@ -91,11 +98,11 @@ class APIError(Exception):
       except json.decoder.JSONDecodeError:
         message = response.text
         response_json = {
-                    "message": message,
-                    "status": response.reason_phrase,
+            "message": message,
+            "status": response.reason_phrase,
         }
     else:
-            response_json = response.body_segments[0].get("error", {})
+      response_json = response.body_segments[0].get("error", {})
 
     status_code = response.status_code
     if 400 <= status_code < 500:
@@ -107,25 +114,55 @@ class APIError(Exception):
 
   @classmethod
   async def raise_for_async_response(
-        cls, response: Union["ReplayResponse", httpx.Response]
+      cls,
+      response: Union[
+          'ReplayResponse', httpx.Response, 'aiohttp.ClientResponse'
+      ],
   ) -> None:
     """Raises an error with detailed error message if the response has an error status."""
-    if response.status_code == 200:
-      return
+    status_code = 0
+    response_json = None
     if isinstance(response, httpx.Response):
+      if response.status_code == 200:
+        return
       try:
         await response.aread()
         response_json = response.json()
       except json.decoder.JSONDecodeError:
         message = response.text
         response_json = {
-                    "message": message,
-                    "status": response.reason_phrase,
+            "message": message,
+            "status": response.reason_phrase,
         }
+      status_code = response.status_code
+    elif hasattr(response, 'body_segments') and hasattr(
+        response, 'status_code'
+    ):
+      if response.status_code == 200:
+        return
+      response_json = response.body_segments[0].get('error', {})
+      status_code = response.status_code
     else:
-            response_json = response.body_segments[0].get("error", {})
+      try:
+        import aiohttp  # pylint: disable=g-import-not-at-top
 
-    status_code = response.status_code
+        if isinstance(response, aiohttp.ClientResponse):
+          if response.status == 200:
+            return
+          try:
+            response_json = await response.json()
+          except aiohttp.client_exceptions.ContentTypeError:
+            message = await response.text()
+            response_json = {
+                'message': message,
+                'status': response.reason,
+            }
+          status_code = response.status
+        else:
+          raise ValueError(f'Unsupported response type: {type(response)}')
+      except ImportError:
+        raise ValueError(f'Unsupported response type: {type(response)}')
+
     if 400 <= status_code < 500:
       raise ClientError(status_code, response_json, response)
     elif 500 <= status_code < 600:
@@ -148,36 +185,39 @@ class ServerError(APIError):
 
 class UnknownFunctionCallArgumentError(ValueError):
   """Raised when the function call argument cannot be converted to the parameter annotation."""
-
   pass
 
 
 class UnsupportedFunctionError(ValueError):
   """Raised when the function is not supported."""
+  pass
 
 
 class FunctionInvocationError(ValueError):
   """Raised when the function cannot be invoked with the given arguments."""
-
   pass
 
 
-class ExperimentalWarning(Warning):
-  """Warning for experimental features."""
+class UnknownApiResponseError(ValueError):
+  """Raised when the response from the API cannot be parsed as JSON."""
+  pass
+
+
+ExperimentalWarning = _common.ExperimentalWarning
 
 
 class CancelledError(ClientError):
-    """The operation was cancelled."""
+  """The operation was cancelled."""
 
 
 class FileProcessingError(Exception):
-    """Error related to file processing in the API.
+  """Error related to file processing in the API.
 
-    This exception is raised when a file fails to reach the ACTIVE state
-    required for using it in content generation requests.
-    """
+  This exception is raised when a file fails to reach the ACTIVE state
+  required for using it in content generation requests.
+  """
 
-    def __init__(self, message, response_json=None):
-        self.message = message
-        self.details = response_json or {}
-        super().__init__(message)
+  def __init__(self, message, response_json=None):
+    self.message = message
+    self.details = response_json or {}
+    super().__init__(message)
