@@ -16,26 +16,89 @@
 # Code generated by the Google Gen AI SDK generator DO NOT EDIT.
 
 import io
+import json
 import logging
-import mimetypes
 import os
-import pathlib
 from typing import Any, Optional, Union
 from urllib.parse import urlencode
+
 from . import _api_module
 from . import _common
+from . import _extra_utils
 from . import _transformers as t
 from . import types
-from ._api_client import BaseApiClient
 from ._common import get_value_by_path as getv
 from ._common import set_value_by_path as setv
 from .pagers import AsyncPager, Pager
 
+
 logger = logging.getLogger('google_genai.files')
 
 
+def _CreateFileParameters_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['file']) is not None:
+    setv(to_object, ['file'], getv(from_object, ['file']))
+
+  return to_object
+
+
+def _CreateFileResponse_from_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['sdkHttpResponse']) is not None:
+    setv(
+        to_object, ['sdk_http_response'], getv(from_object, ['sdkHttpResponse'])
+    )
+
+  return to_object
+
+
+def _DeleteFileParameters_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['name']) is not None:
+    setv(
+        to_object, ['_url', 'file'], t.t_file_name(getv(from_object, ['name']))
+    )
+
+  return to_object
+
+
+def _DeleteFileResponse_from_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['sdkHttpResponse']) is not None:
+    setv(
+        to_object, ['sdk_http_response'], getv(from_object, ['sdkHttpResponse'])
+    )
+
+  return to_object
+
+
+def _GetFileParameters_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['name']) is not None:
+    setv(
+        to_object, ['_url', 'file'], t.t_file_name(getv(from_object, ['name']))
+    )
+
+  return to_object
+
+
 def _ListFilesConfig_to_mldev(
-    api_client: BaseApiClient,
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
@@ -57,270 +120,31 @@ def _ListFilesConfig_to_mldev(
 
 
 def _ListFilesParameters_to_mldev(
-    api_client: BaseApiClient,
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
   if getv(from_object, ['config']) is not None:
-    setv(
-        to_object,
-        ['config'],
-        _ListFilesConfig_to_mldev(
-            api_client, getv(from_object, ['config']), to_object
-        ),
-    )
-
-  return to_object
-
-
-def _FileStatus_to_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['details']) is not None:
-    setv(to_object, ['details'], getv(from_object, ['details']))
-
-  if getv(from_object, ['message']) is not None:
-    setv(to_object, ['message'], getv(from_object, ['message']))
-
-  if getv(from_object, ['code']) is not None:
-    setv(to_object, ['code'], getv(from_object, ['code']))
-
-  return to_object
-
-
-def _File_to_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['name']) is not None:
-    setv(to_object, ['name'], getv(from_object, ['name']))
-
-  if getv(from_object, ['display_name']) is not None:
-    setv(to_object, ['displayName'], getv(from_object, ['display_name']))
-
-  if getv(from_object, ['mime_type']) is not None:
-    setv(to_object, ['mimeType'], getv(from_object, ['mime_type']))
-
-  if getv(from_object, ['size_bytes']) is not None:
-    setv(to_object, ['sizeBytes'], getv(from_object, ['size_bytes']))
-
-  if getv(from_object, ['create_time']) is not None:
-    setv(to_object, ['createTime'], getv(from_object, ['create_time']))
-
-  if getv(from_object, ['expiration_time']) is not None:
-    setv(to_object, ['expirationTime'], getv(from_object, ['expiration_time']))
-
-  if getv(from_object, ['update_time']) is not None:
-    setv(to_object, ['updateTime'], getv(from_object, ['update_time']))
-
-  if getv(from_object, ['sha256_hash']) is not None:
-    setv(to_object, ['sha256Hash'], getv(from_object, ['sha256_hash']))
-
-  if getv(from_object, ['uri']) is not None:
-    setv(to_object, ['uri'], getv(from_object, ['uri']))
-
-  if getv(from_object, ['download_uri']) is not None:
-    setv(to_object, ['downloadUri'], getv(from_object, ['download_uri']))
-
-  if getv(from_object, ['state']) is not None:
-    setv(to_object, ['state'], getv(from_object, ['state']))
-
-  if getv(from_object, ['source']) is not None:
-    setv(to_object, ['source'], getv(from_object, ['source']))
-
-  if getv(from_object, ['video_metadata']) is not None:
-    setv(to_object, ['videoMetadata'], getv(from_object, ['video_metadata']))
-
-  if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _FileStatus_to_mldev(
-            api_client, getv(from_object, ['error']), to_object
-        ),
-    )
-
-  return to_object
-
-
-def _CreateFileParameters_to_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['file']) is not None:
-    setv(
-        to_object,
-        ['file'],
-        _File_to_mldev(api_client, getv(from_object, ['file']), to_object),
-    )
-
-  if getv(from_object, ['config']) is not None:
-    setv(to_object, ['config'], getv(from_object, ['config']))
-
-  return to_object
-
-
-def _GetFileParameters_to_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['name']) is not None:
-    setv(
-        to_object,
-        ['_url', 'file'],
-        t.t_file_name(api_client, getv(from_object, ['name'])),
-    )
-
-  if getv(from_object, ['config']) is not None:
-    setv(to_object, ['config'], getv(from_object, ['config']))
-
-  return to_object
-
-
-def _DeleteFileParameters_to_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['name']) is not None:
-    setv(
-        to_object,
-        ['_url', 'file'],
-        t.t_file_name(api_client, getv(from_object, ['name'])),
-    )
-
-  if getv(from_object, ['config']) is not None:
-    setv(to_object, ['config'], getv(from_object, ['config']))
-
-  return to_object
-
-
-def _FileStatus_from_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['details']) is not None:
-    setv(to_object, ['details'], getv(from_object, ['details']))
-
-  if getv(from_object, ['message']) is not None:
-    setv(to_object, ['message'], getv(from_object, ['message']))
-
-  if getv(from_object, ['code']) is not None:
-    setv(to_object, ['code'], getv(from_object, ['code']))
-
-  return to_object
-
-
-def _File_from_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['name']) is not None:
-    setv(to_object, ['name'], getv(from_object, ['name']))
-
-  if getv(from_object, ['displayName']) is not None:
-    setv(to_object, ['display_name'], getv(from_object, ['displayName']))
-
-  if getv(from_object, ['mimeType']) is not None:
-    setv(to_object, ['mime_type'], getv(from_object, ['mimeType']))
-
-  if getv(from_object, ['sizeBytes']) is not None:
-    setv(to_object, ['size_bytes'], getv(from_object, ['sizeBytes']))
-
-  if getv(from_object, ['createTime']) is not None:
-    setv(to_object, ['create_time'], getv(from_object, ['createTime']))
-
-  if getv(from_object, ['expirationTime']) is not None:
-    setv(to_object, ['expiration_time'], getv(from_object, ['expirationTime']))
-
-  if getv(from_object, ['updateTime']) is not None:
-    setv(to_object, ['update_time'], getv(from_object, ['updateTime']))
-
-  if getv(from_object, ['sha256Hash']) is not None:
-    setv(to_object, ['sha256_hash'], getv(from_object, ['sha256Hash']))
-
-  if getv(from_object, ['uri']) is not None:
-    setv(to_object, ['uri'], getv(from_object, ['uri']))
-
-  if getv(from_object, ['downloadUri']) is not None:
-    setv(to_object, ['download_uri'], getv(from_object, ['downloadUri']))
-
-  if getv(from_object, ['state']) is not None:
-    setv(to_object, ['state'], getv(from_object, ['state']))
-
-  if getv(from_object, ['source']) is not None:
-    setv(to_object, ['source'], getv(from_object, ['source']))
-
-  if getv(from_object, ['videoMetadata']) is not None:
-    setv(to_object, ['video_metadata'], getv(from_object, ['videoMetadata']))
-
-  if getv(from_object, ['error']) is not None:
-    setv(
-        to_object,
-        ['error'],
-        _FileStatus_from_mldev(
-            api_client, getv(from_object, ['error']), to_object
-        ),
-    )
+    _ListFilesConfig_to_mldev(getv(from_object, ['config']), to_object)
 
   return to_object
 
 
 def _ListFilesResponse_from_mldev(
-    api_client: BaseApiClient,
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
+  if getv(from_object, ['sdkHttpResponse']) is not None:
+    setv(
+        to_object, ['sdk_http_response'], getv(from_object, ['sdkHttpResponse'])
+    )
+
   if getv(from_object, ['nextPageToken']) is not None:
     setv(to_object, ['next_page_token'], getv(from_object, ['nextPageToken']))
 
   if getv(from_object, ['files']) is not None:
-    setv(
-        to_object,
-        ['files'],
-        [
-            _File_from_mldev(api_client, item, to_object)
-            for item in getv(from_object, ['files'])
-        ],
-    )
-
-  return to_object
-
-
-def _CreateFileResponse_from_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
-  if getv(from_object, ['httpHeaders']) is not None:
-    setv(to_object, ['http_headers'], getv(from_object, ['httpHeaders']))
-
-  return to_object
-
-
-def _DeleteFileResponse_from_mldev(
-    api_client: BaseApiClient,
-    from_object: Union[dict[str, Any], object],
-    parent_object: Optional[dict[str, Any]] = None,
-) -> dict[str, Any]:
-  to_object: dict[str, Any] = {}
+    setv(to_object, ['files'], [item for item in getv(from_object, ['files'])])
 
   return to_object
 
@@ -357,9 +181,7 @@ class Files(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _ListFilesParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _ListFilesParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files'.format_map(request_url_dict)
@@ -382,17 +204,18 @@ class Files(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = self._api_client.request(
-        'get', path, request_dict, http_options
-    )
+    response = self._api_client.request('get', path, request_dict, http_options)
+
+    response_dict = {} if not response.body else json.loads(response.body)
 
     if not self._api_client.vertexai:
-      response_dict = _ListFilesResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _ListFilesResponse_from_mldev(response_dict)
 
     return_value = types.ListFilesResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
+    )
+    return_value.sdk_http_response = types.HttpResponse(
+        headers=response.headers
     )
     self._api_client._verify_response(return_value)
     return return_value
@@ -414,9 +237,7 @@ class Files(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _CreateFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _CreateFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'upload/v1beta/files'.format_map(request_url_dict)
@@ -439,18 +260,26 @@ class Files(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = self._api_client.request(
+    response = self._api_client.request(
         'post', path, request_dict, http_options
     )
 
+    if config is not None and getattr(
+        config, 'should_return_http_response', None
+    ):
+      return_value = types.CreateFileResponse(sdk_http_response=response)
+      self._api_client._verify_response(return_value)
+      return return_value
+
+    response_dict = {} if not response.body else json.loads(response.body)
+
     if not self._api_client.vertexai:
-      response_dict = _CreateFileResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _CreateFileResponse_from_mldev(response_dict)
 
     return_value = types.CreateFileResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
     )
+
     self._api_client._verify_response(return_value)
     return return_value
 
@@ -485,9 +314,7 @@ class Files(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _GetFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _GetFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files/{file}'.format_map(request_url_dict)
@@ -510,16 +337,14 @@ class Files(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = self._api_client.request(
-        'get', path, request_dict, http_options
-    )
+    response = self._api_client.request('get', path, request_dict, http_options)
 
-    if not self._api_client.vertexai:
-      response_dict = _File_from_mldev(self._api_client, response_dict)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     return_value = types.File._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
     )
+
     self._api_client._verify_response(return_value)
     return return_value
 
@@ -553,9 +378,7 @@ class Files(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _DeleteFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _DeleteFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files/{file}'.format_map(request_url_dict)
@@ -578,17 +401,20 @@ class Files(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = self._api_client.request(
+    response = self._api_client.request(
         'delete', path, request_dict, http_options
     )
 
+    response_dict = {} if not response.body else json.loads(response.body)
+
     if not self._api_client.vertexai:
-      response_dict = _DeleteFileResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _DeleteFileResponse_from_mldev(response_dict)
 
     return_value = types.DeleteFileResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
+    )
+    return_value.sdk_http_response = types.HttpResponse(
+        headers=response.headers
     )
     self._api_client._verify_response(return_value)
     return return_value
@@ -596,7 +422,7 @@ class Files(_api_module.BaseModule):
   def upload(
       self,
       *,
-      file: Union[str, pathlib.Path, os.PathLike[str], io.IOBase],
+      file: Union[str, os.PathLike[str], io.IOBase],
       config: Optional[types.UploadFileConfigOrDict] = None,
   ) -> types.File:
     """Calls the API to upload a file using a supported file service.
@@ -629,79 +455,43 @@ class Files(_api_module.BaseModule):
     if file_obj.name is not None and not file_obj.name.startswith('files/'):
       file_obj.name = f'files/{file_obj.name}'
 
-    if isinstance(file, io.IOBase):
-      if file_obj.mime_type is None:
-        raise ValueError(
-            'Unknown mime type: Could not determine the mimetype for your'
-            ' file\n please set the `mime_type` argument'
-        )
-      if hasattr(file, 'mode'):
-        if 'b' not in file.mode:
-          raise ValueError('The file must be opened in binary mode.')
-      offset = file.tell()
-      file.seek(0, os.SEEK_END)
-      file_obj.size_bytes = file.tell() - offset
-      file.seek(offset, os.SEEK_SET)
-    else:
-      fs_path = os.fspath(file)
-      if not fs_path or not os.path.isfile(fs_path):
-        raise FileNotFoundError(f'{file} is not a valid file path.')
-      file_obj.size_bytes = os.path.getsize(fs_path)
-      if file_obj.mime_type is None:
-        file_obj.mime_type, _ = mimetypes.guess_type(fs_path)
-      if file_obj.mime_type is None:
-        raise ValueError(
-            'Unknown mime type: Could not determine the mimetype for your'
-            ' file\n    please set the `mime_type` argument'
-        )
-
-    http_options: types.HttpOptions
-    if config_model and config_model.http_options:
-      http_options = config_model.http_options
-      http_options.api_version = ''
-      http_options.headers = {
-          'Content-Type': 'application/json',
-          'X-Goog-Upload-Protocol': 'resumable',
-          'X-Goog-Upload-Command': 'start',
-          'X-Goog-Upload-Header-Content-Length': f'{file_obj.size_bytes}',
-          'X-Goog-Upload-Header-Content-Type': f'{file_obj.mime_type}',
-      }
-    else:
-      http_options = types.HttpOptions(
-          api_version='',
-          headers={
-              'Content-Type': 'application/json',
-              'X-Goog-Upload-Protocol': 'resumable',
-              'X-Goog-Upload-Command': 'start',
-              'X-Goog-Upload-Header-Content-Length': f'{file_obj.size_bytes}',
-              'X-Goog-Upload-Header-Content-Type': f'{file_obj.mime_type}',
-          },
-      )
+    http_options, size_bytes, mime_type = _extra_utils.prepare_resumable_upload(
+        file,
+        user_http_options=config_model.http_options,
+        user_mime_type=config_model.mime_type,
+    )
+    file_obj.size_bytes = size_bytes
+    file_obj.mime_type = mime_type
     response = self._create(
-        file=file_obj, config=types.CreateFileConfig(http_options=http_options)
+        file=file_obj,
+        config=types.CreateFileConfig(
+            http_options=http_options, should_return_http_response=True
+        ),
     )
 
     if (
-        response.http_headers is None
-        or 'x-goog-upload-url' not in response.http_headers
+        response.sdk_http_response is None
+        or response.sdk_http_response.headers is None
+        or 'x-goog-upload-url' not in response.sdk_http_response.headers
     ):
       raise KeyError(
           'Failed to create file. Upload URL did not returned from the create'
           ' file request.'
       )
-    upload_url = response.http_headers['x-goog-upload-url']
+    upload_url = response.sdk_http_response.headers['x-goog-upload-url']
 
     if isinstance(file, io.IOBase):
       return_file = self._api_client.upload_file(
           file, upload_url, file_obj.size_bytes, http_options=http_options
       )
     else:
+      fs_path = os.fspath(file)
       return_file = self._api_client.upload_file(
           fs_path, upload_url, file_obj.size_bytes, http_options=http_options
       )
 
     return types.File._from_response(
-        response=_File_from_mldev(self._api_client, return_file.json['file']),
+        response=return_file.json['file'],
         kwargs=config_model.model_dump() if config else {},
     )
 
@@ -774,7 +564,7 @@ class Files(_api_module.BaseModule):
           'downloaded. You can tell which files are downloadable by checking '
           'the `source` or `download_uri` property.'
       )
-    name = t.t_file_name(self._api_client, file)
+    name = t.t_file_name(file)
 
     path = f'files/{name}:download'
 
@@ -829,9 +619,7 @@ class AsyncFiles(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _ListFilesParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _ListFilesParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files'.format_map(request_url_dict)
@@ -854,17 +642,20 @@ class AsyncFiles(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = await self._api_client.async_request(
+    response = await self._api_client.async_request(
         'get', path, request_dict, http_options
     )
 
+    response_dict = {} if not response.body else json.loads(response.body)
+
     if not self._api_client.vertexai:
-      response_dict = _ListFilesResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _ListFilesResponse_from_mldev(response_dict)
 
     return_value = types.ListFilesResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
+    )
+    return_value.sdk_http_response = types.HttpResponse(
+        headers=response.headers
     )
     self._api_client._verify_response(return_value)
     return return_value
@@ -886,9 +677,7 @@ class AsyncFiles(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _CreateFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _CreateFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'upload/v1beta/files'.format_map(request_url_dict)
@@ -911,18 +700,26 @@ class AsyncFiles(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = await self._api_client.async_request(
+    response = await self._api_client.async_request(
         'post', path, request_dict, http_options
     )
 
+    if config is not None and getattr(
+        config, 'should_return_http_response', None
+    ):
+      return_value = types.CreateFileResponse(sdk_http_response=response)
+      self._api_client._verify_response(return_value)
+      return return_value
+
+    response_dict = {} if not response.body else json.loads(response.body)
+
     if not self._api_client.vertexai:
-      response_dict = _CreateFileResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _CreateFileResponse_from_mldev(response_dict)
 
     return_value = types.CreateFileResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
     )
+
     self._api_client._verify_response(return_value)
     return return_value
 
@@ -957,9 +754,7 @@ class AsyncFiles(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _GetFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _GetFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files/{file}'.format_map(request_url_dict)
@@ -982,16 +777,16 @@ class AsyncFiles(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = await self._api_client.async_request(
+    response = await self._api_client.async_request(
         'get', path, request_dict, http_options
     )
 
-    if not self._api_client.vertexai:
-      response_dict = _File_from_mldev(self._api_client, response_dict)
+    response_dict = {} if not response.body else json.loads(response.body)
 
     return_value = types.File._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
     )
+
     self._api_client._verify_response(return_value)
     return return_value
 
@@ -1025,9 +820,7 @@ class AsyncFiles(_api_module.BaseModule):
           'This method is only supported in the Gemini Developer client.'
       )
     else:
-      request_dict = _DeleteFileParameters_to_mldev(
-          self._api_client, parameter_model
-      )
+      request_dict = _DeleteFileParameters_to_mldev(parameter_model)
       request_url_dict = request_dict.get('_url')
       if request_url_dict:
         path = 'files/{file}'.format_map(request_url_dict)
@@ -1050,17 +843,20 @@ class AsyncFiles(_api_module.BaseModule):
     request_dict = _common.convert_to_dict(request_dict)
     request_dict = _common.encode_unserializable_types(request_dict)
 
-    response_dict = await self._api_client.async_request(
+    response = await self._api_client.async_request(
         'delete', path, request_dict, http_options
     )
 
+    response_dict = {} if not response.body else json.loads(response.body)
+
     if not self._api_client.vertexai:
-      response_dict = _DeleteFileResponse_from_mldev(
-          self._api_client, response_dict
-      )
+      response_dict = _DeleteFileResponse_from_mldev(response_dict)
 
     return_value = types.DeleteFileResponse._from_response(
         response=response_dict, kwargs=parameter_model.model_dump()
+    )
+    return_value.sdk_http_response = types.HttpResponse(
+        headers=response.headers
     )
     self._api_client._verify_response(return_value)
     return return_value
@@ -1068,7 +864,7 @@ class AsyncFiles(_api_module.BaseModule):
   async def upload(
       self,
       *,
-      file: Union[str, pathlib.Path, os.PathLike[str], io.IOBase],
+      file: Union[str, os.PathLike[str], io.IOBase],
       config: Optional[types.UploadFileConfigOrDict] = None,
   ) -> types.File:
     """Calls the API to upload a file asynchronously using a supported file service.
@@ -1101,78 +897,48 @@ class AsyncFiles(_api_module.BaseModule):
     if file_obj.name is not None and not file_obj.name.startswith('files/'):
       file_obj.name = f'files/{file_obj.name}'
 
-    if isinstance(file, io.IOBase):
-      if file_obj.mime_type is None:
-        raise ValueError(
-            'Unknown mime type: Could not determine the mimetype for your'
-            ' file\n    please set the `mime_type` argument'
-        )
-      if hasattr(file, 'mode'):
-        if 'b' not in file.mode:
-          raise ValueError('The file must be opened in binary mode.')
-      offset = file.tell()
-      file.seek(0, os.SEEK_END)
-      file_obj.size_bytes = file.tell() - offset
-      file.seek(offset, os.SEEK_SET)
-    else:
-      fs_path = os.fspath(file)
-      if not fs_path or not os.path.isfile(fs_path):
-        raise FileNotFoundError(f'{file} is not a valid file path.')
-      file_obj.size_bytes = os.path.getsize(fs_path)
-      if file_obj.mime_type is None:
-        file_obj.mime_type, _ = mimetypes.guess_type(fs_path)
-      if file_obj.mime_type is None:
-        raise ValueError(
-            'Unknown mime type: Could not determine the mimetype for your'
-            ' file\n    please set the `mime_type` argument'
-        )
-
-    http_options: types.HttpOptions
-    if config_model and config_model.http_options:
-      http_options = config_model.http_options
-      http_options.api_version = ''
-      http_options.headers = {
-          'Content-Type': 'application/json',
-          'X-Goog-Upload-Protocol': 'resumable',
-          'X-Goog-Upload-Command': 'start',
-          'X-Goog-Upload-Header-Content-Length': f'{file_obj.size_bytes}',
-          'X-Goog-Upload-Header-Content-Type': f'{file_obj.mime_type}',
-      }
-    else:
-      http_options = types.HttpOptions(
-          api_version='',
-          headers={
-              'Content-Type': 'application/json',
-              'X-Goog-Upload-Protocol': 'resumable',
-              'X-Goog-Upload-Command': 'start',
-              'X-Goog-Upload-Header-Content-Length': f'{file_obj.size_bytes}',
-              'X-Goog-Upload-Header-Content-Type': f'{file_obj.mime_type}',
-          },
-      )
+    http_options, size_bytes, mime_type = _extra_utils.prepare_resumable_upload(
+        file,
+        user_http_options=config_model.http_options,
+        user_mime_type=config_model.mime_type,
+    )
+    file_obj.size_bytes = size_bytes
+    file_obj.mime_type = mime_type
     response = await self._create(
-        file=file_obj, config=types.CreateFileConfig(http_options=http_options)
+        file=file_obj,
+        config=types.CreateFileConfig(
+            http_options=http_options, should_return_http_response=True
+        ),
     )
     if (
-        response.http_headers is None
-        or 'x-goog-upload-url' not in response.http_headers
+        response.sdk_http_response is None
+        or response.sdk_http_response.headers is None
+        or (
+            'x-goog-upload-url' not in response.sdk_http_response.headers
+            and 'X-Goog-Upload-URL' not in response.sdk_http_response.headers
+        )
     ):
       raise KeyError(
           'Failed to create file. Upload URL did not returned from the create'
           ' file request.'
       )
-    upload_url = response.http_headers['x-goog-upload-url']
+    elif 'x-goog-upload-url' in response.sdk_http_response.headers:
+      upload_url = response.sdk_http_response.headers['x-goog-upload-url']
+    else:
+      upload_url = response.sdk_http_response.headers['X-Goog-Upload-URL']
 
     if isinstance(file, io.IOBase):
       return_file = await self._api_client.async_upload_file(
           file, upload_url, file_obj.size_bytes, http_options=http_options
       )
     else:
+      fs_path = os.fspath(file)
       return_file = await self._api_client.async_upload_file(
           fs_path, upload_url, file_obj.size_bytes, http_options=http_options
       )
 
     return types.File._from_response(
-        response=_File_from_mldev(self._api_client, return_file.json['file']),
+        response=return_file.json['file'],
         kwargs=config_model.model_dump() if config else {},
     )
 
@@ -1233,7 +999,7 @@ class AsyncFiles(_api_module.BaseModule):
       else:
         config_model = config
 
-    name = t.t_file_name(self._api_client, file)
+    name = t.t_file_name(file)
 
     path = f'files/{name}:download'
 
