@@ -18,12 +18,20 @@
 # pylint: disable=protected-access
 
 import copy
-from typing import Any, AsyncIterator,Awaitable, Callable, Generic, Iterator, Literal, TypeVar
+from typing import Any, AsyncIterator, Awaitable, Callable, Generic, Iterator, Literal, TypeVar, Union
+from . import _common
+from . import types
 
 T = TypeVar('T')
 
 PagedItem = Literal[
-    'batch_jobs', 'models', 'tuning_jobs', 'files', 'cached_contents'
+    'batch_jobs',
+    'models',
+    'tuning_jobs',
+    'files',
+    'cached_contents',
+    'file_search_stores',
+    'documents',
 ]
 
 
@@ -42,6 +50,8 @@ class _BasePager(Generic[T]):
 
     self._page = getattr(response, self._name) or []
     self._idx = 0
+
+    self._sdk_http_response = getattr(response, 'sdk_http_response', None)
 
     if not config:
       request_config = {}
@@ -65,7 +75,7 @@ class _BasePager(Generic[T]):
 
   @property
   def page(self) -> list[T]:
-    """Returns a subset of the entire list of items. 
+    """Returns a subset of the entire list of items.
 
     For the number of items returned, see `pageSize()`.
 
@@ -111,7 +121,13 @@ class _BasePager(Generic[T]):
     return self._page_size
 
   @property
-  def config(self) -> dict[str, Any]:
+  def sdk_http_response(self) -> Union[types.HttpResponse, None]:
+    """Returns the http response of the API response."""
+
+    return self._sdk_http_response
+
+  @property
+  def config(self) -> _common.StringDict:
     """Returns the configuration when making the API request for the next page.
 
     A configuration is a set of optional parameters and arguments that can be
