@@ -16,19 +16,41 @@
 
 """Tests for client initialization."""
 
+import asyncio
+import concurrent.futures
 import logging
 import os
 import ssl
+from unittest import mock
 
 import certifi
 import google.auth
 from google.auth import credentials
+import httpx
 import pytest
 
 from ... import _api_client as api_client
 from ... import _base_url as base_url
 from ... import _replay_api_client as replay_api_client
 from ... import Client
+from ... import types
+try:
+  import aiohttp
+  AIOHTTP_NOT_INSTALLED = False
+except ImportError:
+  AIOHTTP_NOT_INSTALLED = True
+  aiohttp = mock.MagicMock()
+
+
+requires_aiohttp = pytest.mark.skipif(
+    AIOHTTP_NOT_INSTALLED, reason="aiohttp is not installed, skipping test."
+)
+
+
+@pytest.fixture(autouse=True)
+def reset_has_aiohttp():
+  yield
+  api_client.has_aiohttp = False
 
 
 def test_ml_dev_from_gemini_env_only(monkeypatch):
@@ -83,6 +105,16 @@ def test_ml_dev_both_env_key_set(monkeypatch, caplog):
       "Both GOOGLE_API_KEY and GEMINI_API_KEY are set. Using GOOGLE_API_KEY."
       in caplog.text
   )
+
+
+def test_api_key_with_new_line(monkeypatch, caplog):
+  caplog.set_level(logging.DEBUG, logger="google_genai._api_client")
+  api_key = "gemini_api_key\r\n"
+  monkeypatch.setenv("GOOGLE_API_KEY", api_key)
+
+  client = Client()
+
+  assert client.models._api_client.api_key == "gemini_api_key"
 
 
 def test_ml_dev_from_constructor():
@@ -206,12 +238,12 @@ def test_constructor_with_invalid_http_options_key():
 
 
 def test_constructor_with_http_options_as_pydantic_type():
-  mldev_http_options = api_client.HttpOptions(
+  mldev_http_options = types.HttpOptions(
       api_version="v1",
       base_url="https://placeholder-fake-url.com/",
       headers={"X-Custom-Header": "custom_value"},
   )
-  vertexai_http_options = api_client.HttpOptions(
+  vertexai_http_options = types.HttpOptions(
       api_version="v1",
       base_url=(
           "https://{self.location}-aiplatform.googleapis.com/{{api_version}}/"
@@ -328,6 +360,21 @@ def test_invalid_vertexai_constructor_empty(monkeypatch):
     Client(vertexai=True)
 
 
+def test_vertexai_constructor_empty_base_url_override(monkeypatch):
+  monkeypatch.setenv("GOOGLE_CLOUD_PROJECT", "")
+  monkeypatch.setenv("GOOGLE_CLOUD_LOCATION", "")
+  monkeypatch.setenv("GOOGLE_API_KEY", "")
+  monkeypatch.setenv("GEMINI_API_KEY", "")
+
+  def mock_auth_default(scopes=None):
+    return None, None
+
+  monkeypatch.setattr(google.auth, "default", mock_auth_default)
+  # Including a base_url override skips the check for having proj/location or
+  # api_key set.
+  Client(vertexai=True, http_options={"base_url": "https://override.com/"})
+
+
 def test_invalid_mldev_constructor_empty(monkeypatch):
   with pytest.raises(ValueError):
     monkeypatch.setenv("GOOGLE_API_KEY", "")
@@ -361,13 +408,120 @@ def test_invalid_vertexai_constructor2():
     )
 
 
-def test_invalid_vertexai_constructor3(monkeypatch):
+def test_vertexai_default_location_to_global(monkeypatch):
 
   with monkeypatch.context() as m:
     m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
     project_id = "fake_project_id"
-    with pytest.raises(ValueError):
-      Client(vertexai=True, project=project_id)
+    client = Client(vertexai=True, project=project_id)
+    assert client.models._api_client.location == "global"
+
+
+def test_vertexai_default_location_to_global_with_credentials(monkeypatch):
+  # Test case 1: When credentials are provided with project but no location
+  creds = credentials.AnonymousCredentials()
+  project_id = "fake_project_id"
+
+  with monkeypatch.context() as m:
+    m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
+    m.setenv("GOOGLE_API_KEY", "")
+    client = Client(vertexai=True, credentials=creds, project=project_id)
+    assert client.models._api_client.location == "global"
+    assert client.models._api_client.project == project_id
+
+
+def test_vertexai_default_location_to_global_with_explicit_project_and_env_apikey(
+    monkeypatch,
+):
+  # Test case 2: When explicit project is provided and env api_key exists
+  project_id = "explicit_project_id"
+  api_key = "env_api_key"
+
+  with monkeypatch.context() as m:
+    m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
+    m.delenv("GOOGLE_CLOUD_PROJECT", raising=False)
+    m.setenv("GOOGLE_API_KEY", api_key)
+    client = Client(vertexai=True, project=project_id)
+    # Explicit project takes precedence over implicit api_key
+    assert client.models._api_client.location == "global"
+    assert client.models._api_client.project == project_id
+    assert not client.models._api_client.api_key
+
+
+def test_vertexai_default_location_to_global_with_env_project_and_env_apikey(
+    monkeypatch,
+):
+  # Test case 3: When env project and env api_key both exist
+  project_id = "env_project_id"
+  api_key = "env_api_key"
+
+  with monkeypatch.context() as m:
+    m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
+    m.setenv("GOOGLE_CLOUD_PROJECT", project_id)
+    m.setenv("GOOGLE_API_KEY", api_key)
+    client = Client(vertexai=True)
+    # Implicit project takes precedence over implicit api_key
+    assert client.models._api_client.location == "global"
+    assert client.models._api_client.project == project_id
+    assert not client.models._api_client.api_key
+
+
+def test_vertexai_no_default_location_when_location_explicitly_set(monkeypatch):
+  # Verify that location is NOT defaulted to global when explicitly set
+  project_id = "fake_project_id"
+  location = "us-central1"
+
+  with monkeypatch.context() as m:
+    m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
+    client = Client(vertexai=True, project=project_id, location=location)
+    assert client.models._api_client.location == location
+    assert client.models._api_client.project == project_id
+
+
+def test_vertexai_no_default_location_when_env_location_set(monkeypatch):
+  # Verify that location is NOT defaulted to global when set via environment
+  project_id = "fake_project_id"
+  location = "us-west1"
+
+  with monkeypatch.context() as m:
+    m.setenv("GOOGLE_CLOUD_LOCATION", location)
+    m.setenv("GOOGLE_CLOUD_PROJECT", project_id)
+    client = Client(vertexai=True)
+    assert client.models._api_client.location == location
+    assert client.models._api_client.project == project_id
+
+
+def test_vertexai_no_default_location_with_apikey_only(monkeypatch):
+  # Verify that location is NOT set when using API key mode (no project)
+  api_key = "vertexai_api_key"
+
+  with monkeypatch.context() as m:
+    m.delenv("GOOGLE_CLOUD_LOCATION", raising=False)
+    m.delenv("GOOGLE_CLOUD_PROJECT", raising=False)
+    m.setenv("GOOGLE_API_KEY", "")
+    client = Client(vertexai=True, api_key=api_key)
+    assert not client.models._api_client.location
+    assert not client.models._api_client.project
+    assert client.models._api_client.api_key == api_key
+
+
+def test_vertexai_explicit_credentials(monkeypatch):
+  creds = credentials.AnonymousCredentials()
+  monkeypatch.setenv("GOOGLE_CLOUD_PROJECT", "fake_project_id")
+  monkeypatch.setenv("GOOGLE_CLOUD_LOCATION", "fake-location")
+  monkeypatch.setenv("GOOGLE_API_KEY", "env_api_key")
+
+  client = Client(
+      vertexai=True,
+      credentials=creds
+  )
+
+  assert client.models._api_client.vertexai
+  assert client.models._api_client.project
+  assert client.models._api_client.location
+  assert not client.models._api_client.api_key
+  assert client.models._api_client._credentials is creds
+  assert isinstance(client.models._api_client, api_client.BaseApiClient)
 
 
 def test_vertexai_explicit_arg_precedence1(monkeypatch):
@@ -489,8 +643,7 @@ def test_change_client_mode_from_env(monkeypatch, use_vertex: bool):
   assert isinstance(
       client1.models._api_client, replay_api_client.ReplayApiClient
   )
-
-  monkeypatch.setenv("GOOGLE_GENAI_CLIENT_MODE", None)
+  monkeypatch.setenv("GOOGLE_GENAI_CLIENT_MODE", "")
 
   client2 = Client()
   assert isinstance(client2.models._api_client, api_client.BaseApiClient)
@@ -557,7 +710,9 @@ def test_vertexai_apikey_from_env_gemini_api_key_only(monkeypatch):
   assert isinstance(client.models._api_client, api_client.BaseApiClient)
 
 
-def test_vertexai_apikey_from_env_gemini_api_key_with_google_api_key_empty(monkeypatch):
+def test_vertexai_apikey_from_env_gemini_api_key_with_google_api_key_empty(
+    monkeypatch,
+):
   # Vertex AI Express mode uses API key on Vertex AI.
   api_key = "vertexai_api_key"
   monkeypatch.setenv("GEMINI_API_KEY", api_key)
@@ -578,7 +733,7 @@ def test_vertexai_apikey_from_env_gemini_api_key_with_google_api_key_empty(monke
   assert isinstance(client.models._api_client, api_client.BaseApiClient)
 
 
-def test_vertexai_apikey_from_env_both_api_keys(monkeypatch,  caplog):
+def test_vertexai_apikey_from_env_both_api_keys(monkeypatch, caplog):
   caplog.set_level(logging.DEBUG, logger="google_genai._api_client")
   # Vertex AI Express mode uses API key on Vertex AI.
   google_api_key = "google_api_key"
@@ -716,14 +871,23 @@ def test_client_logs_to_logger_instance(monkeypatch, caplog):
 
 
 def test_client_ssl_context_implicit_initialization():
-  client_args, async_client_args = api_client.BaseApiClient._ensure_ssl_ctx(
-      api_client.HttpOptions()
+  client_args, async_client_args = (
+      api_client.BaseApiClient._ensure_httpx_ssl_ctx(types.HttpOptions())
   )
 
   assert client_args["verify"]
-  assert async_client_args["verify"]
   assert isinstance(client_args["verify"], ssl.SSLContext)
-  assert isinstance(async_client_args["verify"], ssl.SSLContext)
+  try:
+    import aiohttp  # pylint: disable=g-import-not-at-top
+
+    async_client_args = api_client.BaseApiClient._ensure_aiohttp_ssl_ctx(
+        types.HttpOptions()
+    )
+    assert async_client_args["ssl"]
+    assert isinstance(async_client_args["ssl"], ssl.SSLContext)
+  except ImportError:
+    assert async_client_args["verify"]
+    assert isinstance(async_client_args["verify"], ssl.SSLContext)
 
 
 def test_client_ssl_context_explicit_initialization_same_args():
@@ -732,15 +896,25 @@ def test_client_ssl_context_explicit_initialization_same_args():
       capath=os.environ.get("SSL_CERT_DIR"),
   )
 
-  options = api_client.HttpOptions(
+  options = types.HttpOptions(
       client_args={"verify": ctx}, async_client_args={"verify": ctx}
   )
-  client_args, async_client_args = api_client.BaseApiClient._ensure_ssl_ctx(
-      options
+  client_args, async_client_args = (
+      api_client.BaseApiClient._ensure_httpx_ssl_ctx(options)
   )
 
   assert client_args["verify"] == ctx
-  assert async_client_args["verify"] == ctx
+  try:
+    import aiohttp  # pylint: disable=g-import-not-at-top
+
+    async_client_args = api_client.BaseApiClient._ensure_aiohttp_ssl_ctx(
+        options
+    )
+    assert async_client_args["ssl"]
+    assert isinstance(async_client_args["ssl"], ssl.SSLContext)
+  except ImportError:
+    assert async_client_args["verify"]
+    assert isinstance(async_client_args["verify"], ssl.SSLContext)
 
 
 def test_client_ssl_context_explicit_initialization_separate_args():
@@ -754,15 +928,25 @@ def test_client_ssl_context_explicit_initialization_separate_args():
       capath=os.environ.get("SSL_CERT_DIR"),
   )
 
-  options = api_client.HttpOptions(
+  options = types.HttpOptions(
       client_args={"verify": ctx}, async_client_args={"verify": async_ctx}
   )
-  client_args, async_client_args = api_client.BaseApiClient._ensure_ssl_ctx(
-      options
+  client_args, async_client_args = (
+      api_client.BaseApiClient._ensure_httpx_ssl_ctx(options)
   )
 
   assert client_args["verify"] == ctx
-  assert async_client_args["verify"] == async_ctx
+  try:
+    import aiohttp  # pylint: disable=g-import-not-at-top
+
+    async_client_args = api_client.BaseApiClient._ensure_aiohttp_ssl_ctx(
+        options
+    )
+    assert async_client_args["ssl"]
+    assert isinstance(async_client_args["ssl"], ssl.SSLContext)
+  except ImportError:
+    assert async_client_args["verify"]
+    assert isinstance(async_client_args["verify"], ssl.SSLContext)
 
 
 def test_client_ssl_context_explicit_initialization_sync_args():
@@ -771,13 +955,23 @@ def test_client_ssl_context_explicit_initialization_sync_args():
       capath=os.environ.get("SSL_CERT_DIR"),
   )
 
-  options = api_client.HttpOptions(client_args={"verify": ctx})
-  client_args, async_client_args = api_client.BaseApiClient._ensure_ssl_ctx(
-      options
+  options = types.HttpOptions(client_args={"verify": ctx})
+  client_args, async_client_args = (
+      api_client.BaseApiClient._ensure_httpx_ssl_ctx(options)
   )
 
   assert client_args["verify"] == ctx
-  assert async_client_args["verify"] == ctx
+  try:
+    import aiohttp  # pylint: disable=g-import-not-at-top
+
+    async_client_args = api_client.BaseApiClient._ensure_aiohttp_ssl_ctx(
+        options
+    )
+    assert async_client_args["ssl"]
+    assert isinstance(async_client_args["ssl"], ssl.SSLContext)
+  except ImportError:
+    assert async_client_args["verify"]
+    assert isinstance(async_client_args["verify"], ssl.SSLContext)
 
 
 def test_client_ssl_context_explicit_initialization_async_args():
@@ -786,13 +980,23 @@ def test_client_ssl_context_explicit_initialization_async_args():
       capath=os.environ.get("SSL_CERT_DIR"),
   )
 
-  options = api_client.HttpOptions(async_client_args={"verify": ctx})
-  client_args, async_client_args = api_client.BaseApiClient._ensure_ssl_ctx(
-      options
+  options = types.HttpOptions(async_client_args={"verify": ctx})
+  client_args, async_client_args = (
+      api_client.BaseApiClient._ensure_httpx_ssl_ctx(options)
   )
 
   assert client_args["verify"] == ctx
-  assert async_client_args["verify"] == ctx
+  try:
+    import aiohttp  # pylint: disable=g-import-not-at-top
+
+    async_client_args = api_client.BaseApiClient._ensure_aiohttp_ssl_ctx(
+        options
+    )
+    assert async_client_args["ssl"]
+    assert isinstance(async_client_args["ssl"], ssl.SSLContext)
+  except ImportError:
+    assert async_client_args["verify"]
+    assert isinstance(async_client_args["verify"], ssl.SSLContext)
 
 
 def test_constructor_with_base_url_from_http_options():
@@ -993,3 +1197,489 @@ def test_constructor_with_base_url_from_environment_variables(monkeypatch):
       ]
       == "https://vertex-base-url.com/"
   )
+
+
+def test_async_transport_absence_allows_aiohttp_to_be_used():
+  client = Client(
+      vertexai=True,
+      project="fake_project_id",
+      location="fake-location",
+  )
+
+  api_client.has_aiohttp = False
+  assert not client._api_client._use_aiohttp()
+
+  api_client.has_aiohttp = True
+  assert client._api_client._use_aiohttp()
+
+
+def test_async_async_client_args_without_transport_allows_aiohttp_to_be_used():
+  client = Client(
+      vertexai=True,
+      project="fake_project_id",
+      location="fake-location",
+      http_options=types.HttpOptions(async_client_args={}),
+  )
+
+  api_client.has_aiohttp = False
+  assert not client._api_client._use_aiohttp()
+
+  api_client.has_aiohttp = True
+  assert client._api_client._use_aiohttp()
+
+
+def test_async_transport_forces_httpx_regardless_of_aiohttp_availability():
+
+  client = Client(
+      vertexai=True,
+      project="fake_project_id",
+      location="fake-location",
+      http_options=types.HttpOptions(
+          async_client_args={"transport": httpx.AsyncBaseTransport()}
+      ),
+  )
+
+  api_client.has_aiohttp = False
+  assert not client._api_client._use_aiohttp()
+
+  api_client.has_aiohttp = True
+  assert not client._api_client._use_aiohttp()
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_basic_functionality():
+  """Tests that _get_async_auth_lock returns an asyncio.Lock."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  lock = await client._api_client._get_async_auth_lock()
+  assert isinstance(lock, asyncio.Lock)
+  assert client._api_client._async_auth_lock is lock
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_returns_same_instance():
+  """Tests that multiple calls return the same lock instance."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+  lock1 = await client._api_client._get_async_auth_lock()
+  lock2 = await client._api_client._get_async_auth_lock()
+  lock3 = await client._api_client._get_async_auth_lock()
+  assert lock1 is lock2
+  assert lock2 is lock3
+  assert isinstance(lock1, asyncio.Lock)
+
+
+def test_threaded_generate_content_locking(monkeypatch):
+  """Tests that synchronous API calls are thread-safe."""
+  monkeypatch.delenv("GOOGLE_GENAI_CLIENT_MODE", raising=False)
+  # Mock credentials
+  mock_creds = mock.Mock(spec=credentials.Credentials)
+  mock_creds.token = "initial-token"
+  mock_creds.expired = False
+  mock_creds.quota_project_id = None
+
+  # Mock google.auth.default
+  mock_auth_default = mock.Mock(return_value=(mock_creds, "test-project"))
+  monkeypatch.setattr(google.auth, "default", mock_auth_default)
+
+  # Mock Credentials.refresh
+  def refresh_side_effect(request):
+    mock_creds.token = "refreshed-token"
+    mock_creds.expired = False
+
+  mock_refresh = mock.Mock(side_effect=refresh_side_effect)
+  mock_creds.refresh = mock_refresh
+
+  # Mock the actual request to avoid network calls
+  mock_httpx_response = httpx.Response(
+      status_code=200,
+      headers={},
+      text='{"candidates": [{"content": {"parts": [{"text": "response"}]}}]}',
+  )
+  mock_request = mock.Mock(return_value=mock_httpx_response)
+  monkeypatch.setattr(api_client.SyncHttpxClient, "request", mock_request)
+
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+  # Reset credentials to test initialization to ensure the sync lock is tested.
+  client._api_client._credentials = None
+
+  # 1. Test initial credential loading in multiple threads
+  with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
+    futures = [
+        executor.submit(
+            client.models.generate_content, model="gemini-pro", contents=str(i)
+        )
+        for i in range(10)
+    ]
+    for future in concurrent.futures.as_completed(futures):
+      assert future.result().text == "response"
+
+  mock_auth_default.assert_called_once()
+  mock_refresh.assert_not_called()
+  assert mock_request.call_count == 10
+
+  # 2. Test credential refreshing in multiple threads
+  mock_creds.expired = True
+  with concurrent.futures.ThreadPoolExecutor(max_workers=3) as executor:
+    futures = [
+        executor.submit(
+            client.models.generate_content, model="gemini-pro", contents=str(i)
+        )
+        for i in range(10)
+    ]
+    for future in concurrent.futures.as_completed(futures):
+      assert future.result().text == "response"
+
+  mock_auth_default.assert_called_once()
+  mock_refresh.assert_called_once()
+  assert mock_request.call_count == 20
+
+
+@pytest.mark.asyncio
+async def test_async_access_token_locking(monkeypatch):
+  """Tests that _async_access_token uses locks to prevent race conditions."""
+  # Mock credentials
+  mock_creds = mock.Mock(spec=credentials.Credentials)
+  mock_creds.token = "initial-token"
+  mock_creds.expired = False
+  mock_creds.quota_project_id = None
+
+  # Mock google.auth.default
+  mock_auth_default = mock.Mock(return_value=(mock_creds, "test-project"))
+  monkeypatch.setattr(google.auth, "default", mock_auth_default)
+
+  # Mock Credentials.refresh
+  def refresh_side_effect(request):
+    mock_creds.token = "refreshed-token"
+    mock_creds.expired = False
+
+  mock_refresh = mock.Mock(side_effect=refresh_side_effect)
+  mock_creds.refresh = mock_refresh
+
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+  # Reset credentials to test initialization to ensure the async lock is tested.
+  client._api_client._credentials = None
+
+  # 1. Test initial credential loading
+  # Running them concurrently should result in only one call to load_auth.
+  tokens = await asyncio.gather(
+      client._api_client._async_access_token(),
+      client._api_client._async_access_token(),
+      client._api_client._async_access_token(),
+  )
+
+  assert tokens == ["initial-token", "initial-token", "initial-token"]
+  mock_auth_default.assert_called_once()
+  mock_refresh.assert_not_called()
+
+  # 2. Test credential refreshing
+  # Now the token is "expired", so the next call should refresh it.
+  mock_creds.expired = True
+
+  # Running them concurrently should result in only one call to refresh.
+  tokens = await asyncio.gather(
+      client._api_client._async_access_token(),
+      client._api_client._async_access_token(),
+      client._api_client._async_access_token(),
+  )
+
+  assert tokens == ["refreshed-token", "refreshed-token", "refreshed-token"]
+  # google.auth.default should still have been called only once in total.
+  mock_auth_default.assert_called_once()
+  mock_refresh.assert_called_once()
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_concurrent_access():
+  """Tests that concurrent access to _get_async_auth_lock is thread-safe."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  # Run multiple concurrent calls
+  async def get_lock_task(task_id: int):
+    lock = await client._api_client._get_async_auth_lock()
+    return task_id, id(lock)
+
+  tasks = [get_lock_task(i) for i in range(20)]
+  results = await asyncio.gather(*tasks)
+
+  # All tasks should get the same lock instance
+  lock_ids = [result[1] for result in results]
+  assert all(
+      lock_id == lock_ids[0] for lock_id in lock_ids
+  ), "All tasks should get the same lock instance"
+
+  # All tasks should complete
+  task_ids = [result[0] for result in results]
+  assert sorted(task_ids) == list(range(20)), "All tasks should complete"
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_doesnt_block_other_operations():
+  """Tests that _get_async_auth_lock doesn't interfere with other async operations."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  # Track completion of other async operations
+  completed_operations = []
+
+  async def mock_async_operation(op_id: int):
+    await asyncio.sleep(0.01)  # Small delay to simulate async work
+    completed_operations.append(op_id)
+    return f"operation_{op_id}"
+
+  # Start auth lock requests and other operations simultaneously
+  start_time = asyncio.get_event_loop().time()
+
+  auth_tasks = [client._api_client._get_async_auth_lock() for _ in range(10)]
+  work_tasks = [mock_async_operation(i) for i in range(15)]
+
+  auth_results, work_results = await asyncio.gather(
+      asyncio.gather(*auth_tasks), asyncio.gather(*work_tasks)
+  )
+
+  end_time = asyncio.get_event_loop().time()
+  total_time = end_time - start_time
+
+  # Verify all operations completed
+  assert len(auth_results) == 10, "All auth lock requests should complete"
+  assert len(work_results) == 15, "All work tasks should complete"
+  assert len(completed_operations) == 15, "All async operations should complete"
+
+  # All auth requests should return the same lock
+  lock_ids = [id(lock) for lock in auth_results]
+  assert all(lock_id == lock_ids[0] for lock_id in lock_ids)
+
+  # Should complete quickly since operations run concurrently
+  assert total_time < 0.1, (
+      f"Operations took too long ({total_time:.3f}s), suggesting blocking"
+      " occurred"
+  )
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_creation_lock_lifecycle():
+  """Tests the creation lock lifecycle and cleanup."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  # Initially, both locks should be None
+  assert client._api_client._async_auth_lock is None
+  assert client._api_client._async_auth_lock_creation_lock is None
+
+  # After first call, both should exist
+  lock1 = await client._api_client._get_async_auth_lock()
+  assert client._api_client._async_auth_lock is not None
+  assert client._api_client._async_auth_lock_creation_lock is not None
+  assert isinstance(lock1, asyncio.Lock)
+
+  # Creation lock should be different from the auth lock
+  creation_lock = client._api_client._async_auth_lock_creation_lock
+  assert creation_lock is not lock1
+  assert isinstance(creation_lock, asyncio.Lock)
+
+  # Subsequent calls should reuse both locks
+  lock2 = await client._api_client._get_async_auth_lock()
+  assert lock2 is lock1
+  assert client._api_client._async_auth_lock_creation_lock is creation_lock
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_under_load():
+  """Tests _get_async_auth_lock under heavy concurrent load."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  num_concurrent_calls = 100
+
+  async def get_lock_with_timing(call_id: int):
+    start = asyncio.get_event_loop().time()
+    lock = await client._api_client._get_async_auth_lock()
+    end = asyncio.get_event_loop().time()
+    return call_id, id(lock), end - start
+
+  # Run many concurrent calls
+  start_time = asyncio.get_event_loop().time()
+  tasks = [get_lock_with_timing(i) for i in range(num_concurrent_calls)]
+  results = await asyncio.gather(*tasks)
+  total_time = asyncio.get_event_loop().time() - start_time
+
+  # Verify all calls succeeded and got the same lock
+  call_ids = [r[0] for r in results]
+  lock_ids = [r[1] for r in results]
+  call_times = [r[2] for r in results]
+
+  assert len(results) == num_concurrent_calls
+  assert sorted(call_ids) == list(range(num_concurrent_calls))
+  assert all(
+      lock_id == lock_ids[0] for lock_id in lock_ids
+  ), "All calls should get same lock"
+
+  # Performance checks
+  max_call_time = max(call_times)
+  assert total_time < 1.0, f"Total time ({total_time:.3f}s) suggests blocking"
+  assert (
+      max_call_time < 0.1
+  ), f"Max individual call time ({max_call_time:.3f}s) too high"
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_interleaved_with_auth_operations():
+  """Tests _get_async_auth_lock working correctly with actual auth operations."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  # Mock credentials for this test
+  mock_creds = mock.Mock(spec=credentials.Credentials)
+  mock_creds.token = "test-token"
+  mock_creds.expired = False
+  mock_creds.quota_project_id = None
+  client._api_client._credentials = mock_creds
+
+  # Mix lock requests with simulated auth operations
+  async def auth_operation(op_id: int):
+    # This simulates what _async_access_token does
+    lock = await client._api_client._get_async_auth_lock()
+    async with lock:
+      await asyncio.sleep(0.001)  # Simulate auth work
+      return f"auth_op_{op_id}"
+
+  async def lock_request(req_id: int):
+    lock = await client._api_client._get_async_auth_lock()
+    return req_id, id(lock)
+
+  # Interleave different types of operations
+  auth_tasks = [auth_operation(i) for i in range(10)]
+  lock_tasks = [lock_request(i) for i in range(10)]
+
+  auth_results, lock_results = await asyncio.gather(
+      asyncio.gather(*auth_tasks), asyncio.gather(*lock_tasks)
+  )
+
+  # Verify all operations completed
+  assert len(auth_results) == 10
+  assert len(lock_results) == 10
+
+  # All lock requests should return the same lock ID
+  lock_ids = [result[1] for result in lock_results]
+  assert all(lock_id == lock_ids[0] for lock_id in lock_ids)
+
+  # Auth operations should complete successfully
+  assert all(result.startswith("auth_op_") for result in auth_results)
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_with_event_loop_switch():
+  """Tests that _get_async_auth_lock works correctly with event loop context."""
+
+  async def create_client_and_get_lock():
+    client = Client(
+        vertexai=True, project="fake_project_id", location="fake-location"
+    )
+    lock = await client._api_client._get_async_auth_lock()
+    return client, lock
+
+  # Create client and get lock in current event loop
+  client, lock1 = await create_client_and_get_lock()
+
+  # Get lock again in same event loop
+  lock2 = await client._api_client._get_async_auth_lock()
+
+  assert lock1 is lock2
+  assert isinstance(lock1, asyncio.Lock)
+
+  # Verify the locks work correctly
+  async def test_lock_functionality():
+    async with lock1:
+      await asyncio.sleep(0.001)
+      return "success"
+
+  result = await test_lock_functionality()
+  assert result == "success"
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_double_checked_locking():
+  """Tests the double-checked locking pattern implementation."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+
+  original_lock_init = asyncio.Lock.__init__
+  lock_creation_count = [0]
+
+  def counting_lock_init(self):
+    lock_creation_count[0] += 1
+    return original_lock_init(self)
+
+  # Patch asyncio.Lock to count creations
+  asyncio.Lock.__init__ = counting_lock_init
+
+  try:
+    # Run many concurrent requests
+    tasks = [client._api_client._get_async_auth_lock() for _ in range(50)]
+    locks = await asyncio.gather(*tasks)
+
+    # All should be the same instance
+    assert all(lock is locks[0] for lock in locks)
+
+    # Should only create 2 locks: creation_lock + auth_lock
+    # (Could be slightly more due to asyncio internals, but should be minimal)
+    assert (
+        lock_creation_count[0] <= 5
+    ), f"Created {lock_creation_count[0]} locks, expected ~2"
+
+  finally:
+    asyncio.Lock.__init__ = original_lock_init
+
+
+@pytest.mark.asyncio
+async def test_get_async_auth_lock_memory_efficiency():
+  """Tests that _get_async_auth_lock doesn't leak memory under repeated use."""
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+  initial_lock = await client._api_client._get_async_auth_lock()
+  initial_creation_lock = client._api_client._async_auth_lock_creation_lock
+
+  # Run many operations
+  for _ in range(100):
+    lock = await client._api_client._get_async_auth_lock()
+    assert lock is initial_lock
+    assert (
+        client._api_client._async_auth_lock_creation_lock
+        is initial_creation_lock
+    )
+  # Verify no new objects were created
+  final_lock = await client._api_client._get_async_auth_lock()
+  final_creation_lock = client._api_client._async_auth_lock_creation_lock
+
+  assert final_lock is initial_lock
+  assert final_creation_lock is initial_creation_lock
+
+
+@requires_aiohttp
+@pytest.mark.asyncio
+async def test_get_aiohttp_session():
+  """Tests that _get_async_auth_lock works correctly with aiohttp session lock."""
+
+  client = Client(
+      vertexai=True, project="fake_project_id", location="fake-location"
+  )
+  api_client.has_aiohttp = True
+  initial_session = await client._api_client._get_aiohttp_session()
+  assert initial_session is not None
+  session = await client._api_client._get_aiohttp_session()
+  assert session is initial_session
