@@ -16,10 +16,9 @@
 
 """Tests for client behavior when issuing requests."""
 
-import http
-
 from ... import _api_client as api_client
 from ... import Client
+from ... import types
 
 
 def build_test_client(monkeypatch):
@@ -31,7 +30,7 @@ def test_join_url_path_base_url_with_trailing_slash_and_path_with_leading_slash(
   base_url = 'https://fake-url.com/some_path/'
   path = '/v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/some_path/v1beta/models'
   )
 
@@ -40,7 +39,7 @@ def test_join_url_path_with_base_url_with_trailing_slash_and_path_without_leadin
   base_url = 'https://fake-url.com/some_path/'
   path = 'v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/some_path/v1beta/models'
   )
 
@@ -49,7 +48,7 @@ def test_join_url_path_with_base_url_without_trailing_slash_and_path_with_leadin
   base_url = 'https://fake-url.com/some_path'
   path = '/v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/some_path/v1beta/models'
   )
 
@@ -58,7 +57,7 @@ def test_join_url_path_with_base_url_without_trailing_slash_and_path_without_lea
   base_url = 'https://fake-url.com/some_path'
   path = 'v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/some_path/v1beta/models'
   )
 
@@ -67,7 +66,7 @@ def test_join_url_path_base_url_without_path_with_trailing_slash():
   base_url = 'https://fake-url.com/'
   path = 'v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/v1beta/models'
   )
 
@@ -76,7 +75,7 @@ def test_join_url_path_base_url_without_path_without_trailing_slash():
   base_url = 'https://fake-url.com'
   path = 'v1beta/models'
   assert (
-      api_client._join_url_path(base_url, path)
+      api_client.join_url_path(base_url, path)
       == 'https://fake-url.com/v1beta/models'
   )
 
@@ -96,7 +95,7 @@ def test_build_request_appends_to_user_agent_headers(monkeypatch):
       'GET',
       'test/path',
       {'key': 'value'},
-      api_client.HttpOptionsDict(
+      types.HttpOptionsDict(
           base_url='test/url',
           api_version='1',
           headers={'user-agent': 'test-user-agent'},
@@ -114,7 +113,7 @@ def test_build_request_appends_to_goog_api_client_headers(monkeypatch):
       'GET',
       'test/path',
       {'key': 'value'},
-      api_client.HttpOptionsDict(
+      types.HttpOptionsDict(
           base_url='test/url',
           api_version='1',
           headers={'x-goog-api-client': 'test-goog-api-client'},
@@ -128,14 +127,14 @@ def test_build_request_appends_to_goog_api_client_headers(monkeypatch):
 
 def test_build_request_keeps_sdk_version_headers(monkeypatch):
   headers_to_inject = {}
-  api_client._append_library_version_headers(headers_to_inject)
+  api_client.append_library_version_headers(headers_to_inject)
   assert 'google-genai-sdk/' in headers_to_inject['user-agent']
   request_client = build_test_client(monkeypatch).models._api_client
   request = request_client._build_request(
       'GET',
       'test/path',
       {'key': 'value'},
-      api_client.HttpOptionsDict(
+      types.HttpOptionsDict(
           base_url='test/url',
           api_version='1',
           headers=headers_to_inject,
@@ -145,3 +144,29 @@ def test_build_request_keeps_sdk_version_headers(monkeypatch):
   assert 'gl-python/' in request.headers['x-goog-api-client']
   assert 'google-genai-sdk/' in request.headers['x-goog-api-client']
   assert 'gl-python/' in request.headers['x-goog-api-client']
+
+
+def build_test_client_no_env_vars(monkeypatch):
+  monkeypatch.delenv('GOOGLE_API_KEY', raising=False)
+  monkeypatch.delenv('GEMINI_API_KEY', raising=False)
+  monkeypatch.delenv('GOOGLE_CLOUD_PROJECT', raising=False)
+  monkeypatch.delenv('GOOGLE_CLOUD_LOCATION', raising=False)
+  return Client(
+      vertexai=True,
+      http_options=types.HttpOptionsDict(
+          base_url='https://custom-base-url.com',
+          headers={'Authorization': 'Bearer fake_access_token'},
+      ),
+    )
+
+
+def test_build_request_with_custom_base_url_no_env_vars(monkeypatch):
+  request_client = (
+      build_test_client_no_env_vars(monkeypatch).models._api_client
+  )
+  request = request_client._build_request(
+      'GET',
+      'test/path',
+      {'key': 'value'},
+  )
+  assert request.url == 'https://custom-base-url.com'
