@@ -23,7 +23,7 @@ from .. import pytest_helper
 test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_delete',
-        parameters=types._DeleteFileParameters(name='files/1vx8znuf0yje'),
+        parameters=types._DeleteFileParameters(name='files/1g583ke2xdsn'),
         exception_if_vertex='only supported in the Gemini Developer client',
         skip_in_api_mode=(
             'The files have a TTL, they cannot be reliably retrieved for a long'
@@ -43,4 +43,4 @@ pytestmark = pytest_helper.setup(
 @pytest.mark.asyncio
 async def test_async(client):
   with pytest_helper.exception_if_vertex(client, ValueError):
-    file = await client.aio.files.get(name='files/vjvu9fwk2qj8')
+    file = await client.aio.files.get(name='files/n1gls7dyh90q')
