@@ -15,23 +15,17 @@
 
 
 """Tests for live.py."""
+import base64
 import json
-import os
 from unittest import mock
 
-import PIL.Image
 import pytest
 from websockets import client
 
+from .. import pytest_helper
 from ... import client as gl_client
 from ... import live
 from ... import types
-
-
-IMAGE_FILE_PATH = os.path.abspath(
-    os.path.join(os.path.dirname(__file__), '../data/google.jpg')
-)
-image = PIL.Image.open(IMAGE_FILE_PATH)
 
 
 def mock_api_client(vertexai=False):
@@ -98,6 +92,27 @@ async def test_send_content_content(mock_websocket, vertexai):
   assert 'client_content' in sent_data
 
   assert sent_data['client_content']['turns'][0]['parts'][0]['text'] == 'test'
+
+
+@pytest.mark.parametrize('vertexai', [True, False])
+@pytest.mark.asyncio
+async def test_send_content_with_blob(mock_websocket, vertexai):
+  session = live.AsyncSession(
+      api_client=mock_api_client(vertexai=vertexai), websocket=mock_websocket
+  )
+  content = types.Content.model_validate(
+      {'parts': [{'inline_data': {'data': b'test'}}]}
+  )
+
+  await session.send_client_content(turns=content)
+  mock_websocket.send.assert_called_once()
+  sent_data = json.loads(mock_websocket.send.call_args[0][0])
+  assert 'client_content' in sent_data
+
+  assert pytest_helper.get_value_ignore_key_case(
+      sent_data['client_content']['turns'][0]['parts'][0], 'inline_data') == {
+          'data': base64.b64encode(b'test').decode()
+      }
 
 
 @pytest.mark.parametrize('vertexai', [True, False])
