@@ -17,9 +17,8 @@ import base64
 import enum
 import os
 
-import PIL.Image
-from pydantic import BaseModel, ValidationError, Field
-from typing import Literal, List, Optional, Union
+from pydantic import BaseModel, ValidationError, Field, ConfigDict
+from typing import Literal, List, Optional, Union, Set
 from datetime import datetime
 import pytest
 import json
@@ -34,7 +33,6 @@ from enum import Enum
 IMAGE_PNG_FILE_PATH = os.path.abspath(
     os.path.join(os.path.dirname(__file__), '../data/google.png')
 )
-image_png = PIL.Image.open(IMAGE_PNG_FILE_PATH)
 
 with open(IMAGE_PNG_FILE_PATH, 'rb') as image_file:
   image_bytes = image_file.read()
@@ -68,7 +66,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_http_options_in_method',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash-002',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
             config={
                 'http_options': test_http_options,
             },
@@ -130,7 +128,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_llama',
         parameters=types._GenerateContentParameters(
             model='meta/llama-3.2-90b-vision-instruct-maas',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
         ),
         exception_if_mldev='404',
         skip_in_api_mode='it will encounter 403 for api mode',
@@ -139,10 +137,10 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_system_instructions',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash',
-            contents=t.t_contents(None, 'high'),
+            contents=t.t_contents('high'),
             config={
                 'system_instruction': t.t_content(
-                    None, 'I say high, you say low'
+                    'I say high, you say low'
                 )
             },
         ),
@@ -152,7 +150,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         exception_if_mldev='not supported',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash-002',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
             config={
                 'labels': {'label1': 'value1', 'label2': 'value2'},
             },
@@ -162,7 +160,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_simple_shared_generation_config',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash-002',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
             config={
                 'max_output_tokens': 100,
                 'top_k': 2,
@@ -179,7 +177,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_2_candidates_gemini_1_5_flash',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash',
-            contents=t.t_contents(None, 'Tell me a story in 30 words.'),
+            contents=t.t_contents('Tell me a story in 30 words.'),
             config={
                 'candidate_count': 2,
             },
@@ -189,7 +187,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_safety_settings_on_difference',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
             config={
                 'safety_settings': safety_settings_with_method,
             },
@@ -200,7 +198,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_penalty',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash-002',
-            contents=t.t_contents(None, 'Tell me a story in 30 words.'),
+            contents=t.t_contents('Tell me a story in 30 words.'),
             config={
                 'presence_penalty': 0.5,
                 'frequency_penalty': 0.5,
@@ -211,7 +209,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_penalty_gemini_1_5_flash',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash',
-            contents=t.t_contents(None, 'Tell me a story in 30 words.'),
+            contents=t.t_contents('Tell me a story in 30 words.'),
             config={
                 'presence_penalty': 0.5,
                 'frequency_penalty': 0.5,
@@ -222,9 +220,27 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_google_search_tool',
         parameters=types._GenerateContentParameters(
             model='gemini-2.0-flash-exp',
-            contents=t.t_contents(None, 'Why is the sky blue?'),
+            contents=t.t_contents('Why is the sky blue?'),
             config=types.GenerateContentConfig(
                 tools=[types.Tool(google_search=types.GoogleSearch())]
+            ),
+        ),
+    ),
+    pytest_helper.TestTableItem(
+        name='test_google_maps_tool',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.0-flash-exp',
+            contents=t.t_contents('Find restaurants near me.'),
+            config=types.GenerateContentConfig(
+                tools=[{'google_maps': {}}],
+                tool_config={
+                    "retrieval_config": {
+                        "lat_lng": {
+                            "latitude": 37.421993,
+                            "longitude": -122.079725,
+                        }
+                    }
+                }
             ),
         ),
     ),
@@ -232,7 +248,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_google_search_tool_with_time_range_filter',
         parameters=types._GenerateContentParameters(
             model='gemini-2.0-flash-exp',
-            contents=t.t_contents(None, 'What is the QQQ stock price?'),
+            contents=t.t_contents('What is the QQQ stock price?'),
             config=types.GenerateContentConfig(
                 tools=[
                     types.Tool(
@@ -252,11 +268,94 @@ test_table: list[pytest_helper.TestTableItem] = [
         ),
     ),
     pytest_helper.TestTableItem(
+        name='test_google_search_tool_with_exclude_domains',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('Why is the sky blue?'),
+            config=types.GenerateContentConfig(
+                tools=[
+                    types.Tool(
+                        google_search=types.GoogleSearch(
+                            exclude_domains=['amazon.com', 'facebook.com']
+                        )
+                    )
+                ]
+            ),
+        ),
+        exception_if_mldev='not supported in',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_google_search_tool_with_blocking_confidence',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('Why is the sky blue?'),
+            config=types.GenerateContentConfig(
+                tools=[
+                    types.Tool(
+                        google_search=types.GoogleSearch(
+                            blocking_confidence=types.PhishBlockThreshold.BLOCK_LOW_AND_ABOVE,
+                        )
+                    )
+                ]
+            ),
+        ),
+        exception_if_mldev='not supported in',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_enterprise_web_search_tool',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('Why is the sky blue?'),
+            config=types.GenerateContentConfig(
+                tools=[
+                    types.Tool(
+                        enterprise_web_search=types.EnterpriseWebSearch()
+                    )
+                ]
+            ),
+        ),
+        exception_if_mldev='not supported in',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_enterprise_web_search_tool_with_exclude_domains',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('Why is the sky blue?'),
+            config=types.GenerateContentConfig(
+                tools=[
+                    types.Tool(
+                        enterprise_web_search=types.EnterpriseWebSearch(
+                            exclude_domains=['amazon.com', 'facebook.com']
+                        )
+                    )
+                ]
+            ),
+        ),
+        exception_if_mldev='not supported in',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_enterprise_web_search_tool_with_blocking_confidence',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('Why is the sky blue?'),
+            config=types.GenerateContentConfig(
+                tools=[
+                    types.Tool(
+                        enterprise_web_search=types.EnterpriseWebSearch(
+                            blocking_confidence=types.PhishBlockThreshold.BLOCK_LOW_AND_ABOVE,
+                        )
+                    )
+                ]
+            ),
+        ),
+        exception_if_mldev='not supported in',
+    ),
+    pytest_helper.TestTableItem(
         name='test_speech_with_config',
         parameters=types._GenerateContentParameters(
             model='gemini-2.0-flash-exp',
             contents=t.t_contents(
-                None, 'Produce a speech response saying "Cheese"'
+                'Produce a speech response saying "Cheese"'
             ),
             config=types.GenerateContentConfig(
                 response_modalities=['audio'],
@@ -276,7 +375,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         parameters=types._GenerateContentParameters(
             model='gemini-2.0-flash-exp',
             contents=t.t_contents(
-                None, 'Alice says "Hi", Bob replies with "what\'s up"?'
+                'Alice says "Hi", Bob replies with "what\'s up"?'
             ),
             config=types.GenerateContentConfig(
                 response_modalities=['audio'],
@@ -312,7 +411,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         parameters=types._GenerateContentParameters(
             model='gemini-2.0-flash-exp',
             contents=t.t_contents(
-                None, 'Alice says "Hi", Bob replies with "what\'s up"?'
+                'Alice says "Hi", Bob replies with "what\'s up"?'
             ),
             config=types.GenerateContentConfig(
                 response_modalities=['audio'],
@@ -387,7 +486,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_response_schema_with_default',
         parameters=types._GenerateContentParameters(
             model='gemini-1.5-flash',
-            contents=t.t_contents(None, 'What is your name?'),
+            contents=t.t_contents('What is your name?'),
             config={
                 'response_mime_type': 'application/json',
                 'response_schema': {
@@ -403,6 +502,33 @@ test_table: list[pytest_helper.TestTableItem] = [
         ),
         ignore_keys=['parsed'],
     ),
+    pytest_helper.TestTableItem(
+        name='test_invalid_model_parameter_path',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash/../../upload/v1beta/files',
+            contents=t.t_contents('What is your name?'),
+        ),
+        exception_if_vertex='invalid model parameter',
+        exception_if_mldev='invalid model parameter',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_invalid_model_parameter_question_mark',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash?',
+            contents=t.t_contents('What is your name?'),
+        ),
+        exception_if_vertex='invalid model parameter',
+        exception_if_mldev='invalid model parameter',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_invalid_model_parameter_ampersand',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash&',
+            contents=t.t_contents('What is your name?'),
+        ),
+        exception_if_vertex='invalid model parameter',
+        exception_if_mldev='invalid model parameter',
+    ),
 ]
 
 pytestmark = pytest_helper.setup(
@@ -414,6 +540,31 @@ pytestmark = pytest_helper.setup(
 pytest_plugins = ('pytest_asyncio',)
 
 
+def test_sync_with_headers(client):
+  response = client.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='Tell me a story in 300 words.',
+  )
+  assert response.sdk_http_response.headers is not None
+  assert response.sdk_http_response.body is None
+
+
+def test_sync_with_full_response(client):
+  response = client.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='Tell me a story in 300 words.',
+      config={
+          'should_return_http_response': True,
+      },
+  )
+  print(response.sdk_http_response.body)
+  assert response.sdk_http_response.headers is not None
+  assert response.sdk_http_response.body is not None
+  assert 'candidates' in response.sdk_http_response.body
+  assert 'content' in response.sdk_http_response.body
+  assert 'parts' in response.sdk_http_response.body
+  assert 'usageMetadata' in response.sdk_http_response.body
+
 @pytest.mark.asyncio
 async def test_async(client):
   response = await client.aio.models.generate_content(
@@ -424,6 +575,33 @@ async def test_async(client):
       },
   )
   assert response.text
+
+
+@pytest.mark.asyncio
+async def test_async_with_headers(client):
+  response = await client.aio.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='Tell me a story in 300 words.',
+  )
+  assert response.sdk_http_response.headers is not None
+  assert response.sdk_http_response.body is None
+
+
+@pytest.mark.asyncio
+async def test_async_with_full_response(client):
+  response = await client.aio.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='Tell me a story in 300 words.',
+      config={
+          'should_return_http_response': True,
+      },
+  )
+  assert response.sdk_http_response.headers is not None
+  assert response.sdk_http_response.body is not None
+  assert 'candidates' in response.sdk_http_response.body
+  assert 'content' in response.sdk_http_response.body
+  assert 'parts' in response.sdk_http_response.body
+  assert 'usageMetadata' in response.sdk_http_response.body
 
 
 def test_sync_stream(client):
@@ -442,6 +620,44 @@ def test_sync_stream(client):
   assert chunks > 2
 
 
+def test_sync_stream_with_should_return_http_headers(client):
+  response = client.models.generate_content_stream(
+      model='gemini-1.5-flash',
+      contents='Tell me a story in 300 words.',
+      config={
+          'http_options': test_http_options,
+      },
+  )
+  chunks = 0
+  for part in response:
+    chunks += 1
+    assert part.text is not None or part.candidates[0].finish_reason
+    assert part.sdk_http_response.headers is not None
+  assert chunks > 2
+
+
+def test_sync_stream_with_non_text_modality(client):
+  response = client.models.generate_content_stream(
+      model='gemini-2.0-flash-preview-image-generation',
+      contents=(
+          'Generate an image of the Eiffel tower with fireworks in the'
+          ' background.'
+      ),
+      config={
+          'response_modalities': ['IMAGE', 'TEXT'],
+      },
+  )
+  chunks = 0
+  for chunk in response:
+    chunks += 1
+    if chunk.candidates[0].finish_reason is not None:
+      continue
+    for part in chunk.parts:
+      assert part.text is not None or part.inline_data is not None
+
+  assert chunks > 2
+
+
 @pytest.mark.asyncio
 async def test_async_stream(client):
   chunks = 0
@@ -453,6 +669,44 @@ async def test_async_stream(client):
   ):
     chunks += 1
     assert part.text is not None or part.candidates[0].finish_reason
+
+  assert chunks > 2
+
+
+@pytest.mark.asyncio
+async def test_async_stream_with_headers(client):
+  chunks = 0
+  async for part in await client.aio.models.generate_content_stream(
+      model='gemini-1.5-flash', contents='Tell me a story in 300 words.',
+      config={
+          'http_options': test_http_options,
+      },
+  ):
+    chunks += 1
+    assert part.text is not None or part.candidates[0].finish_reason
+    assert part.sdk_http_response.headers is not None
+
+  assert chunks > 2
+
+
+@pytest.mark.asyncio
+async def test_async_stream_with_non_text_modality(client):
+  chunks = 0
+  async for chunk in await client.aio.models.generate_content_stream(
+      model='gemini-2.0-flash-preview-image-generation',
+      contents=(
+          'Generate an image of the Eiffel tower with fireworks in the'
+          ' background.'
+      ),
+      config={
+          'response_modalities': ['IMAGE', 'TEXT'],
+      },
+  ):
+    chunks += 1
+    if chunk.candidates[0].finish_reason is not None:
+      continue
+    for part in chunk.parts:
+      assert part.text is not None or part.inline_data is not None
 
   assert chunks > 2
 
@@ -518,10 +772,6 @@ async def test_simple_shared_generation_config_stream_async(client):
 
 
 def test_log_probs(client):
-  # ML DEV discovery doc supports response_logprobs but the backend
-  # does not.
-  # TODO: update replay test json files when ML Dev backend is updated.
-  with pytest_helper.exception_if_mldev(client, errors.ClientError):
     client.models.generate_content(
         model='gemini-1.5-flash',
         contents='What is your name?',
@@ -713,6 +963,34 @@ def test_pydantic_schema(client):
       },
   )
   assert isinstance(response.parsed, CountryInfo)
+
+def test_json_schema_fields(client):
+  class UserRole(str, Enum):
+    ADMIN = "admin"
+    VIEWER = "viewer"
+  class Address(BaseModel):
+    street: str
+    city: str
+  class UserProfile(BaseModel):
+    username: str = Field(description="User's unique name")
+    age: Optional[int] = Field(ge=0, le=20)
+    roles: Set[UserRole] = Field(min_items=1)
+    contact: Union[Address, str]
+
+    model_config = ConfigDict(
+        title="User Schema", description="A user profile"
+    )  # This is the title of the schema
+
+  response = client.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='Give me information of the United States.',
+      config={
+          'response_mime_type': 'application/json',
+          'response_json_schema': UserProfile.model_json_schema(),
+      },
+  )
+  print(response.parsed)
+  assert response.parsed != None
 
 
 def test_pydantic_schema_orders_properties(client):
@@ -1458,17 +1736,18 @@ def test_non_string_enum_schema_with_enum_mime_type(client):
     BRASS = 4
     KEYBOARD = 5
 
-  with pytest.raises(TypeError) as e:
-    client.models.generate_content(
-        model='gemini-1.5-flash',
-        contents='What instrument plays multiple notes at once?',
-        config={
-            'response_mime_type': 'text/x.enum',
-            'response_schema': IntegerEnum,
-        },
-    )
+  response =client.models.generate_content(
+      model='gemini-1.5-flash',
+      contents='What instrument plays multiple notes at once?',
+      config={
+          'response_mime_type': 'text/x.enum',
+          'response_schema': IntegerEnum,
+      },
+  )
 
-    assert 'value must be a string' in str(e)
+  instrument_values = {str(member.value) for member in IntegerEnum}
+
+  assert response.text in instrument_values
 
 
 def test_json_schema(client):
@@ -1690,7 +1969,7 @@ def test_json_schema_with_streaming(client):
   )
 
   for r in response:
-    parts = r.candidates[0].content.parts
+    parts = r.parts
     for p in parts:
       assert p.text
 
@@ -1716,7 +1995,7 @@ def test_pydantic_schema_with_streaming(client):
   )
 
   for r in response:
-    parts = r.candidates[0].content.parts
+    parts = r.parts
     for p in parts:
       assert p.text
 
@@ -1807,7 +2086,6 @@ def test_function(client):
       },
   )
   assert '100' in response.text
-
 
 def test_invalid_input_without_transformer(client):
   with pytest.raises(ValidationError) as e:
@@ -2023,33 +2301,6 @@ def test_usage_metadata_part_types(client):
   assert modalities == ['IMAGE', 'TEXT']
 
 
-def test_warning_log_includes_parsed_for_multi_candidate_response(client, caplog):
-  caplog.set_level(logging.DEBUG, logger='google_genai')
-
-  class CountryInfo(BaseModel):
-    name: str
-    population: int
-    capital: str
-    continent: str
-    major_cities: list[str]
-    gdp: int
-    official_language: str
-    total_area_sq_mi: int
-
-  response = client.models.generate_content(
-      model='gemini-2.0-flash',
-      contents='Give me information of the United States.',
-      config={
-          'response_mime_type': 'application/json',
-          'response_schema': CountryInfo,
-          'candidate_count': 2,
-      },
-  )
-  assert response.parsed
-  assert len(response.candidates) == 2
-  assert 'parsed' in caplog.text
-
-
 def test_error_handling_stream(client):
   if client.vertexai:
     return
@@ -2193,8 +2444,4 @@ async def test_error_handling_stream_async(client):
       continue
 
   except errors.ClientError as e:
-    assert (
-        e.message
-        == 'Developer instruction is not enabled for'
-        ' models/gemini-2.0-flash-exp-image-generation'
-    )
+    assert ('Developer instruction is not enabled' in e.message)
