@@ -25,23 +25,13 @@ tuned_model_endpoint = (
     'projects/964831358985/locations/us-central1/endpoints/7226683110069370880'
 )
 
-tuned_model_with_model_name = 'tunedModels/generate-num-8498'
 
 test_table: list[pytest_helper.TestTableItem] = [
-    pytest_helper.TestTableItem(
-        name='test_tuned_model_with_model_name',
-        parameters=types._GenerateContentParameters(
-            model=tuned_model_with_model_name,
-            contents=t.t_contents(None, 'how are you doing?'),
-        ),
-        exception_if_vertex='404',
-        skip_in_api_mode='It requires a specific ML Dev account.',
-    ),
     pytest_helper.TestTableItem(
         name='test_tuned_model',
         parameters=types._GenerateContentParameters(
             model=tuned_model_endpoint,
-            contents=t.t_contents(None, 'Tell me a story in 300 words.'),
+            contents=t.t_contents('Tell me a story in 300 words.'),
         ),
         exception_if_mldev='404',
     ),
@@ -49,7 +39,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_start_with_publishers',
         parameters=types._GenerateContentParameters(
             model='publishers/google/models/gemini-1.5-flash',
-            contents=t.t_contents(None, 'Tell me a story in 50 words.'),
+            contents=t.t_contents('Tell me a story in 50 words.'),
         ),
         exception_if_mldev='404',
     ),
@@ -57,7 +47,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_start_with_models',
         parameters=types._GenerateContentParameters(
             model='models/gemini-1.5-flash',
-            contents=t.t_contents(None, 'Tell me a story in 50 words.'),
+            contents=t.t_contents('Tell me a story in 50 words.'),
         ),
         exception_if_vertex='404',
     ),
@@ -65,7 +55,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_publisher_model',
         parameters=types._GenerateContentParameters(
             model='google/gemini-1.5-flash',
-            contents=t.t_contents(None, 'Tell me a story in 50 words.'),
+            contents=t.t_contents('Tell me a story in 50 words.'),
         ),
         exception_if_mldev='404',
     ),
@@ -73,7 +63,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_empty_model',
         parameters=types._GenerateContentParameters(
             model='',
-            contents=t.t_contents(None, 'Tell me a story in 50 words.'),
+            contents=t.t_contents('Tell me a story in 50 words.'),
         ),
         exception_if_mldev='model',
         exception_if_vertex='model',
@@ -95,18 +85,6 @@ def test_tuned_model_stream(client):
     chunks = 0
     for chunk in client.models.generate_content_stream(
         model=tuned_model_endpoint,
-        contents='Tell me a story in 300 words.',
-    ):
-      chunks += 1
-      assert chunk.text is not None or chunk.candidates[0].finish_reason
-    assert chunks >= 2
-
-
-def test_tuned_model_with_model_name_stream(client):
-  with pytest_helper.exception_if_vertex(client, errors.ClientError):
-    chunks = 0
-    for chunk in client.models.generate_content_stream(
-        model=tuned_model_with_model_name,
         contents='Tell me a story in 300 words.',
     ):
       chunks += 1
