@@ -22,7 +22,7 @@ from ... import types
 from .. import pytest_helper
 
 
-_MODEL = 'models/gemini-2.0-flash-live-001'
+_MODEL = 'models/gemini-live-2.5-flash-preview'
 
 
 # All tests will be run for both Vertex and MLDev.
@@ -30,7 +30,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_no_lock',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={'http_options': {'api_version': 'v1alpha'}},
         ),
         exception_if_vertex='only supported in the Gemini Developer client',
@@ -38,7 +38,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_global_lock',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'uses': 2,
@@ -56,7 +56,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_lock_non_null_fields',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'uses': 2,
@@ -75,7 +75,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_lock_unset_fields_as_default',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'uses': 2,
@@ -94,7 +94,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_lock_additional_fields',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'uses': 2,
@@ -113,7 +113,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_lock_with_no_params',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'lock_additional_fields': ['output_audio_transcription'],
@@ -124,7 +124,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_create_lock_with_empty_params',
         parameters=types.CreateAuthTokenParameters(
-            # remove v1alpha1 after v1beta is available.
+            # remove v1alpha after v1beta is available.
             config={
                 'http_options': {'api_version': 'v1alpha'},
                 'lock_additional_fields': ['output_audio_transcription'],
