@@ -24,35 +24,35 @@ from ... import types
 
 def test_none():
   with pytest.raises(ValueError):
-    t.t_content(None, None)
+    t.t_content( None)
 
 
 def test_content():
   assert t.t_content(
-      None, types.Content(parts=[types.Part(text='test')])
+      types.Content(parts=[types.Part(text='test')])
   ) == types.Content(parts=[types.Part(text='test')])
 
 
 def test_content_dict():
   assert t.t_content(
-      None, {'role': 'user', 'parts': [{'text': 'test'}]}
+      {'role': 'user', 'parts': [{'text': 'test'}]}
   ) == types.Content(parts=[types.Part(text='test')], role='user')
 
 
 def test_content_dict_invalid():
   with pytest.raises(pydantic.ValidationError):
-    t.t_content(None, {'invalid_key': 'test'})
+    t.t_content({'invalid_key': 'test'})
 
 
 def test_text_part_dict():
-  assert t.t_content(None, {'text': 'test'}) == types.UserContent(
+  assert t.t_content({'text': 'test'}) == types.UserContent(
       parts=[types.Part(text='test')]
   )
 
 
 def test_function_call_part_dict():
   assert t.t_content(
-      None, {'function_call': {'name': 'test_func', 'args': {'arg1': 'value1'}}}
+      {'function_call': {'name': 'test_func', 'args': {'arg1': 'value1'}}}
   ) == (
       types.ModelContent(
           parts=[
@@ -67,14 +67,13 @@ def test_function_call_part_dict():
 
 
 def test_text_part():
-  assert t.t_content(None, types.Part(text='test')) == types.UserContent(
+  assert t.t_content(types.Part(text='test')) == types.UserContent(
       parts=[types.Part(text='test')]
   )
 
 
 def test_function_call_part():
   assert t.t_content(
-      None,
       types.Part(
           function_call=types.FunctionCall(
               name='test_func', args={'arg1': 'value1'}
@@ -94,14 +93,14 @@ def test_function_call_part():
 
 
 def test_string():
-  assert t.t_content(None, 'test') == types.UserContent(
+  assert t.t_content('test') == types.UserContent(
       parts=[types.Part(text='test')]
   )
 
 
 def test_file():
   assert t.t_content(
-      None, types.File(uri='gs://test', mime_type='image/png')
+      types.File(uri='gs://test', mime_type='image/png')
   ) == types.UserContent(
       parts=[
           types.Part(
@@ -115,16 +114,47 @@ def test_file():
 
 def test_file_no_uri():
   with pytest.raises(ValueError):
-    t.t_content(None, types.File(mime_type='image/png'))
+    t.t_content(types.File(mime_type='image/png'))
 
 
 def test_file_no_mime_type():
   with pytest.raises(ValueError):
-    t.t_content(None, types.File(uri='gs://test'))
+    t.t_content(types.File(uri='gs://test'))
 
+
+def test_file_dict():
+  assert t.t_content({'file_uri': 'gs://test', 'mime_type': 'image/png'}) == types.UserContent(
+          parts=[
+              types.Part(
+                  file_data=types.FileData(
+                      file_uri='gs://test', mime_type='image/png'
+                  )
+              )
+          ]
+      )
 
 def test_int():
   try:
-    t.t_content(None, 1)
+    t.t_content(1)
   except ValueError as e:
     assert 'Unsupported content part type: <class \'int\'>' in str(e)
+
+
+@pytest.mark.parametrize(
+    'contents',
+    [
+        types.Content(parts=[types.Part(inline_data=types.Blob(data=b'test'))]),
+        {'parts': [{'inline_data': {'data': b'test'}}]},
+        [
+            types.Content(
+                parts=[types.Part(inline_data=types.Blob(data=b'test'))]
+            )
+        ],
+        [{'parts': [{'inline_data': {'data': b'test'}}]}],
+    ],
+)
+def test_t_contents_strict_content(contents):
+  result = t.t_contents_strict(contents)
+  assert result == [
+      types.Content(parts=[types.Part(inline_data=types.Blob(data=b'test'))])
+  ]
